+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # 0.1.0
 
+## New Components
+
+* Alert
+
+## Dialog Type Components
+
+  Breaking Change! :warning: Both components now require a `cancelHandler` prop (rather than the `cancelDialogHandler`). :warning:
+
 ## Dropdowns
 
 Dropdown components have been refactored. We now have three different kinds:
@@ -25,7 +33,7 @@ It is still recommended that you use names on inputs, as they are useful to iden
 ## Dropdown and DropdownSuggest components
 
 Dropdown and dropdown-suggest have been updated. As they share common functionality, dropdown and dropdown-suggest now use a List decorator. This should not affect how you use either component.
-* Dropdown now filters results as you type. 
+* Dropdown now filters results as you type.
 
 ## Dialog Cancel button